requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
python-multipart>=0.0.6
openai>=1.0.0
httpx>=0.24.0
